edit data type amount #8

+++ app/model/customer_expenses.go
@@ -2,7 +2,7 @@ package model
 
 type CustomerExpenses struct {
 	Title  string   `db:"picture" json:"picture"`
-	Amount int      `db:"amount" json:"amount"`
+	Amount float64  `db:"amount" json:"amount"`
 	Note   string   `db:"note" json:"note"`
 	Tags   []string `db:"tags" json:"tags"`
 }
@@ -13,14 +13,14 @@ func NewCusExpenses() *CustomerExpenses {
 
 type CreateCustomerExpense struct {
 	Title  string   `json:"picture"`
-	Amount int      `json:"amount"`
+	Amount float64  `json:"amount"`
 	Note   string   `json:"note"`
 	Tags   []string `json:"tags"`
 }
 
 type UpdateCustomerExpense struct {
 	Title  string   `json:"picture"`
-	Amount int      `json:"amount"`
+	Amount float64  `json:"amount"`
 	Note   string   `json:"note"`
 	Tags   []string `json:"tags"`
 }
